Fixed float to int conversion warning
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@2962 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/CWorm.cpp
@@ -696,7 +696,7 @@ void CWorm::Draw(SDL_Surface * bmpDest, CViewport *v)
 	{
 		int damageDrawPos = WormNameY + tLX->cFont.GetHeight(); 
 		// Make it float up a bit when time passes
-		damageDrawPos += ( tLX->fCurTime - DamageReportDrawOrder.begin()->first ) * 30;
+		damageDrawPos += (int)(( tLX->fCurTime - DamageReportDrawOrder.begin()->first ) * 30);
 		for( std::map< float, int > :: iterator it = DamageReportDrawOrder.begin();
 				it != DamageReportDrawOrder.end(); it++ )
 		{
